requirements.txt
numpy
pandas
scikit-learn
scikit-image
matplotlib
seaborn
pytest
python-dotenv
